+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Fixed internal server error when acessing rate-limited endpoints via unix socket.
 - Fixed incorrect rate limiter IP address check.
+- Fixed incorrect de-duplication of activity recipients.
 
 ## [3.22.0] - 2025-04-02